git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@712 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/Gorgon/Gorgon.Common/Native/Win32API.cs
@@ -76,9 +76,9 @@ namespace GorgonLibrary.Native
 		/// <summary>Message to process.</summary>
 		public uint Message;
 		/// <summary>Window message parameter 1.</summary>
-		public uint wParam;
+		public IntPtr wParam;
 		/// <summary>Window message parameter 2.</summary>
-		public uint lParam;
+		public IntPtr lParam;
 		/// <summary>Time message was sent?</summary>
 		public uint time;
 		/// <summary>Mouse pointer position.</summary>
